requirements.txt
flask-restful
gunicorn
sqlalchemy
psycopg2-binary
marshmallow
flask-apispec
alembic
sendgrid
facebook-sdk
flask-cors
names
googlemaps
firebase-admin
